Revert "Create MomHomeScreen"
This reverts commit 5a2eb1acc3038b57f3f591e467afc420acb52159.
